min/250926 자재요청 ing

+++ server/scm/src/main/java/com/yedam/scm/vo/PrdPlanDetailVO.java
@@ -1,4 +1,3 @@
-
 /*
  * 2025.09.22
  * 생산계획 세부VO [detail]
@@ -18,8 +17,8 @@ import lombok.Data;
 public class PrdPlanDetailVO {
 
     private Long plDetId;      // PK seq부여
-    private String prodId;     // 제품 ID -> 모달
-    private Integer proQty;    // 수량 
+    private String prodId;     // 제품 ID 
+    private Long proQty;    // 수량 
     private Date proDate;      // 생산일자
     private Long plId;         // 계획 ID (FK)
     private String matStatus;  // 자재 상태
